refactor separate tables to new components
