+++ src/components/people.component.tsx
@@ -4,35 +4,25 @@ import {
   FormControl,
   InputGroup,
   Row,
-  Table,
   Col,
 } from 'react-bootstrap';
-import {
-  BsCheck, BsPencil, BsTrash, BsX,
-} from 'react-icons/bs';
 import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 import { Person } from '../models/person.model';
 import '../assets/style/table.css';
 import { ApplicationState, Actions } from '../store';
-import { Expense } from '../models/expense.model';
+import PeopleTableComponent from './people.table.component';
 
 interface StateProps {
   people: Person[];
-  expenses: Expense[];
 }
 
 interface State {
   name: string;
-  editingPerson?: Person;
-  editingName: string;
 }
 
 interface DispatchProps {
   createPersonRequest(data: { state: Person[]; data: Person }): void;
-  editPersonRequest(data: { state: Person[]; data: Person }): void;
-  removePersonRequest(data: { state: Person[]; data: Person }): void;
-  removeExpenseRequest(data: { state: Expense[]; data: Expense }): void;
   loadRequest(): void;
 }
 
@@ -44,13 +34,9 @@ class PeopleComponent extends Component<Props, State> {
 
     this.state = {
       name: '',
-      editingPerson: undefined,
-      editingName: '',
     };
 
     this.updateName = this.updateName.bind(this);
-    this.updateEditingPerson = this.updateEditingPerson.bind(this);
-    this.updateEditingName = this.updateEditingName.bind(this);
   }
 
   componentDidMount() {
@@ -63,24 +49,12 @@ class PeopleComponent extends Component<Props, State> {
     this.setState({ name: event.target.value });
   }
 
-  updateEditingName(event: React.ChangeEvent<HTMLInputElement>) {
-    this.setState({ editingName: event.target.value });
-  }
-
-  updateEditingPerson(person?: Person) {
-    this.setState({ editingPerson: person });
-  }
-
   render() {
     const {
       people,
-      expenses,
       createPersonRequest,
-      editPersonRequest,
-      removePersonRequest,
-      removeExpenseRequest,
     } = this.props;
-    const { name, editingPerson, editingName } = this.state;
+    const { name } = this.state;
 
     return (
       <div className="person-component">
@@ -121,78 +95,7 @@ class PeopleComponent extends Component<Props, State> {
 
         <hr />
 
-        <Table striped bordered hover responsive variant="dark">
-          <thead>
-            <tr>
-              <th>Nome</th>
-              <th className="actions-col">Ações</th>
-            </tr>
-          </thead>
-          <tbody>
-            {people.map((person) => (
-              <tr key={person.id}>
-                {editingPerson === person
-                  ? (
-                    <>
-                      <td>
-                        <InputGroup size="sm">
-                          <InputGroup.Text id="new-name">Novo Nome</InputGroup.Text>
-                          <FormControl onChange={this.updateEditingName} />
-                        </InputGroup>
-                      </td>
-                      <td className="actions-col">
-                        <Button
-                          variant="light"
-                          onClick={() => {
-                            editPersonRequest({
-                              state: people,
-                              data: { ...person, name: editingName },
-                            });
-                            this.updateEditingPerson(undefined);
-                          }}
-                        >
-                          <BsCheck />
-                        </Button>
-                        <Button variant="light" onClick={() => this.updateEditingPerson(undefined)}>
-                          <BsX />
-                        </Button>
-                      </td>
-                    </>
-                  )
-                  : (
-                    <>
-                      <td>
-                        {person.name}
-                      </td>
-                      <td className="actions-col">
-                        <Button variant="light" onClick={() => this.updateEditingPerson(person)}>
-                          <BsPencil />
-                        </Button>
-                        <Button
-                          variant="light"
-                          onClick={() => {
-                            removePersonRequest({
-                              state: people,
-                              data: person,
-                            });
-                            const expense = expenses.find((e) => e.person === person);
-                            if (expense) {
-                              removeExpenseRequest({
-                                state: expenses,
-                                data: expense,
-                              });
-                            }
-                          }}
-                        >
-                          <BsTrash />
-                        </Button>
-                      </td>
-                    </>
-                  )}
-              </tr>
-            ))}
-          </tbody>
-        </Table>
+        <PeopleTableComponent />
       </div>
     );
   }
@@ -200,7 +103,6 @@ class PeopleComponent extends Component<Props, State> {
 
 const mapStateToProps = (state: ApplicationState) => ({
   people: state.people.data,
-  expenses: state.expense.data,
 });
 
 const mapDispatchToProps = (dispatch: Dispatch) => bindActionCreators(Actions, dispatch);
